Fix old ssids offset
Do not encode the old ssid offset in the partition table. Instead,
detect the flash size and set the appropriate offset as there are both
2MB and 4MB modules in the wild.

The older firmware seems to put the old ssids in an address beyond 2MB.
On 4MB modules this is not an issue. But on 2MB modules, accesses beyond
the flash size seems to wrap around. Do not rely on this behavior, and
instead detect the chip size to set the proper offset.